requirements.txt
apipkg==1.5
appdirs==1.4.4
attrs==20.2.0
base58==2.0.1
bitarray==1.2.2
black==19.10b0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cytoolz==0.10.1
eth-abi==2.1.1
eth-account==0.5.2
eth-brownie==1.11.4
eth-event==1.2.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.0
eth-typing==2.2.2
eth-utils==1.9.5
execnet==1.7.1
hexbytes==0.2.1
hypothesis==5.35.0
idna==2.10
inflection==0.5.0
iniconfig==1.0.1
ipfshttpclient==0.6.1
jsonschema==3.2.0
lru-dict==1.1.6
more-itertools==8.5.0
multiaddr==0.0.9
mythx-models==1.9.1
netaddr==0.8.0
packaging==20.4
parsimonious==0.8.1
pathspec==0.8.0
pkg-resources==0.0.0
pluggy==0.13.1
prompt-toolkit==3.0.7
protobuf==3.13.0
psutil==5.7.2
py==1.9.0
py-solc-ast==1.2.5
py-solc-x==1.0.0
pycryptodome==3.9.8
Pygments==2.6.1
pygments-lexer-solidity==0.5.1
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.0.1
pytest-forked==1.3.0
pytest-xdist==1.34.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pythx==1.6.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
rlp==1.2.0
semantic-version==2.8.5
six==1.15.0
sortedcontainers==2.2.2
toml==0.10.1
toolz==0.10.0
tqdm==4.48.2
typed-ast==1.4.1
urllib3==1.25.10
varint==1.0.2
vvm==0.0.2
wcwidth==0.2.5
web3==5.11.1
websockets==8.1
